change resouce type name & add Aws::DynamoDB::Table
