+++ doc/resource_types.md
@@ -11,7 +11,7 @@
 | [cloudwatch_logs](#cloudwatch_logs)
 | [customer_gateway](#customer_gateway)
 | [directconnect_virtual_interface](#directconnect_virtual_interface)
-| [dynamodb](#dynamodb)
+| [dynamodb_table](#dynamodb_table)
 | [ebs](#ebs)
 | [ec2](#ec2)
 | [ecr_repository](#ecr_repository)
@@ -2295,39 +2295,17 @@ end
 ```
 
 ### its(:log_group_name), its(:creation_time), its(:retention_in_days), its(:metric_filter_count), its(:arn), its(:stored_bytes)
-## <a name="dynamodb">dynamodb</a>
+## <a name="dynamodb_table">dynamodb_table</a>
 
-Dynamodb resource type.
+DynamodbTable resource type.
 
 ### exist
-
-```ruby
-describe dynamodb('my-dynamodb-table') do
-  it { should exist }
-end
-```
-
 
 ### be_creating, be_updating, be_deleting, be_active
 
 ### have_attribute_definition
 
-```ruby
-describe dynamodb('my-dynamodb-table') do
-  it { should have_attribute_definition('my-dynamodb-table-attaribute1').attribute_type('S') }
-  it { should have_attribute_definition('my-dynamodb-table-attaribute2').attribute_type('N') }
-end
-```
-
-
 ### have_key_schema
-
-```ruby
-describe dynamodb('my-dynamodb-table') do
-  it { should have_key_schema('my-dynamodb-table-key_schema1').key_type('HASH') }
-  it { should have_key_schema('my-dynamodb-table-key_schema2').key_type('RANGE') }
-end
-```
 
 ### its(:table_name), its(:table_status), its(:creation_date_time), its(:table_size_bytes), its(:item_count), its(:table_arn), its(:local_secondary_indexes), its(:global_secondary_indexes), its(:stream_specification), its(:latest_stream_label), its(:latest_stream_arn)
 # Account and Attributes
