+++ lib/awspec/stub/dynamodb_table.rb
