+++ spec/type/dynamodb_table_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
-Awspec::Stub.load 'dynamodb'
+Awspec::Stub.load 'dynamodb_table'
 
-describe dynamodb('my-dynamodb-table') do
+describe dynamodb_table('my-dynamodb-table') do
   it { should exist }
   it { should be_active }
   its('provisioned_throughput.read_capacity_units') { should eq 1 }
@@ -10,4 +10,10 @@ describe dynamodb('my-dynamodb-table') do
   it { should have_attribute_definition('my-dynamodb-table-attaribute2').attribute_type('N') }
   it { should have_key_schema('my-dynamodb-table-key_schema1').key_type('HASH') }
   it { should have_key_schema('my-dynamodb-table-key_schema2').key_type('RANGE') }
+  context 'nested attribute call' do
+    its(:resource) { should be_an_instance_of(Awspec::ResourceReader) }
+    its('resource.name') { should eq 'my-dynamodb-table' }
+    its('provisioned_throughput.read_capacity_units') { should eq 1 }
+    its('provisioned_throughput.write_capacity_units') { should eq 1 }
+  end
 end
